Move `formatLocalTime` to `string` subdirectory
